bitsperity-homegrow/automation-engine/requirements.txt
websockets==10.4
paho-mqtt==2.1.0
pymongo==4.3.3
python-dotenv==1.0.0
colorlog==6.7.0
pyyaml==6.0 